requirements.txt
celery==5.4.0
Django==5.1.3
djangorestframework==3.15.2
redis==5.2.0
